Add a generator for Starlark builtin wrappers.
The generator looks for functions tagged with the comment
`//starlark:builtin` and generates a Starlark wrapper guided by the
contents of the comment block. The comment block must contain a single
Starlark function definition with an optional docstring. The wrapper
calls `starlark.UnpackArgs` using the names of the Starlark parameters
and the types of the marked function's parameters. Parameters with the
default value "None" are considered optional.

In the near future, the generator will also generate Markdown API
documentation.