Add MVS-based dependency management
These changes add dependency management for dawn projects based on
[minimal version selection]. This results in a Go-style experience for
dawn projects and their dependencies. For the time being, dependencies
are limited to Git projects stored in git repositories. Dependencies are
automatically installed when a project is loaded. Dependencies are
managed using the `dawn get` and `dawn tidy` comamnds. The former
accepts roughly the same arguments as `go get`, including version
queries and the `-u` flag.