Limit runner parallelism
These changes limit parallelism in the runner to runtime.NumCPU. This is
necessary to avoid creating an overwhelming number of threads when
dealing with large projects (especially projects with a large number of
source files).

Limited parallelism is managed by gating runnability on available
capacity. In order to avoid deadlocks, a target releases its capacity
while it is waiting on its dependencies.