chore: update doc comment format
Change the doc comment format to cooperate better with new gofmt rules.
Doc comments are now processed by scraping out the contents of all code
blocks. In order to ensure that code blocks are properly processed, the
first line of the containing Go doc comment must not be blank.

For example, this:

// def foo():
//     """
//     foo does this!
//     """
//
//starlark:builtin

Becomes this:

// starlark
//
//	def foo():
//	    """
//	    foo does this!
//	    """
//
//starlark:builtin

This keeps gofmt from reformatting the code.